pyproject.toml
[tool.poetry]
name = "my-business"
version = "0.1.0"
description = ""
authors = ["Valeriya <Dzhavadova.v@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"
pydantic-settings = "^2.5.2"
sqlalchemy = "^2.0.35"
alembic = "^1.13.3"
asyncpg = "^0.29.0"
pydantic = { extras = ["email"], version = "^2.9.2" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
redis = "^5.1.1"
python-multipart = "^0.0.12"
sqlalchemy-utils = "^0.41.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
exclude = ["migrations/*"]
target-version = "py311"

[tool.ruff.lint]
select = ["ALL"]
extend-select = ["W", "E"]
preview = true
ignore = [
    "ARG001",  # unused-function-argument
    "ANN401",  # any-type
    "B008",    # function-call-in-default-argument
    "B904",    # raise-without-from-inside-except
    "BLE001",  # blind-except
    "CPY001",  # missing-copyright-notice
    "D100",    # undocumented-public-module
    "D101",    # undocumented-public-class
    "D102",    # undocumented-public-method
    "D103",    # undocumented-public-function
    "D104",    # undocumented-public-package
    "D105",    # undocumented-magic-method
    "D107",    # undocumented-public-init
    "D203",    # one-blank-line-before-class
    "D205",    # blank-line-after-summary
    "D213",    # multi-line-summary-second-line
    "D401",    # non-imperative-mood
    "DOC501",  # docstring-missing-exception
    "FAST002", # fast-api-non-annotated-dependency
    "G004",    # logging-f-string
    "INP001",  # implicit-namespace-package
    "RUF009",  # function-call-in-dataclass-default-argument
    "S101",    # assert
    "S608",    # hardcoded-sql-expression
    "TRY003",  # raise-vanilla-args
    "TRY401",  # verbose-log-message
]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.pylint]
max-args = 7
